api: workaround broken 'maxworkers' alias in bulk actions api calls
It seems, that 'alias' schema does not properly work currently for
normal schema definitions (only for property-string), since it does not
verification of the actual schema (checked by calling the api with
maxworkers=1000, which throws no errors). Luckily, this is the only use
of alias for api parameters that I could find. For the new cluster wide
api calls it did not do anything, since we did not actually use
`$param->{maxworkers}`, but for the node specific one broke the UI,
since there we still send 'maxworkers', but the backend did not use that
anymore.

In the long term, we should fix it in JSONSchema, but as a stop-gap,
restore the old 'maxworkers' parameters to a 'real' property and mark it
deprecated.

Also actually use it in the code and fix the error message to the new
variant.

Reported in the forums for the node-specific one:
https://forum.proxmox.com/threads/bulk-migrate-error-in-9-1-2.177326/

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.kernel.org/r/20251205100317.1101549-1-d.csapak@proxmox.com