api: apt: restrict package name and separate arguments from options
In order to sanitize the package names - restrict them to a smaller
set, in addition to the set allowed by the policy-manual:
https://www.debian.org/doc/debian-policy/ch-controlfields.html#s-f-package
I added ':' used as a architecture separator (e.g. zlib1g:amd64).

Follows the commit in pmg-api addressing the issue.

Difference to PMG is that the call here needs 'Sys.Modify', which is
quite a high privilege - in all predefined roles it is only present in
the 'Administrator'.

Fixes: e337b294 ("apt: use `apt changelog` for changelog fetching")
Reported-by: Kevin Joensen <kevin@baldur.dk>
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>